requirements.txt
django==3.2.10
django-cors-headers==3.7.0
django-ranged-response==0.2.0
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
django-filter==2.4.0
pytz==2021.1
whitenoise==5.2.0
